[♻️REFACTOR]  : newChatItem mutation 공용 hook 으로 분리
  - gcTime 적용
