+++ src/components/chat/ChatMessages.tsx
@@ -4,7 +4,6 @@ import { Fragment, useRef, ElementRef, useEffect, useState } from "react";
 import { Message } from "@prisma/client";
 import { Loader2, ServerCrash } from "lucide-react";
 import { format } from "date-fns";
-import { toast } from "sonner";
 
 import useChatQuery from "@/hooks/useChatQuery";
 import useChatSocket from "@/hooks/useChatSocket";
@@ -13,9 +12,9 @@ import useChatScroll from "@/hooks/useChatScroll";
 import ChatWelcome from "./ChatWelcome";
 import ChatItem from "./ChatItem";
 import { useMutationState } from "@tanstack/react-query";
-import { NewChatProps } from "./ChatInput";
 import useAuthStore from "@/store/useAuthsStore";
 import { useSessionStorage } from "usehooks-ts";
+import { NewChatProps } from "@/hooks/useNewChatMutation";
 
 const DATE_FORMAT = "d MMM yyyy, HH:mm";
